backend/requirements.txt
Django==5.1.6
asgiref==3.8.1
django-mongoengine==0.5.6
djangorestframework==3.15.2
dnspython==2.7.0
mongoengine==0.29.1
pip==25.0.1
ruff==0.11.1
sqlparse==0.5.3
tzdata==2025.1
cfgv==3.4.0
colorama==0.4.6
distlib==0.3.9
filelock==3.18.0
identify==2.6.9
iniconfig==2.1.0
nodeenv==1.9.1
packaging==24.2
parameterized==0.9.0
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
pymongo==4.11.1
pytest==8.3.5
pytest-django==4.11.1
PyYAML==6.0.2
virtualenv==20.30.0
